refactor(items): split god-service into 5 domain sub-services
